+++ opusfilter/filters.py
@@ -8,6 +8,7 @@ import os
 import string
 from typing import Iterator, List, Tuple
 
+from iso639 import Lang
 import regex
 
 from . import FilterABC, ConfigurationError, CLEAN_LOW, CLEAN_HIGH, CLEAN_BETWEEN, CLEAN_TRUE, CLEAN_FALSE
@@ -297,7 +298,7 @@ class LanguageIDFilter(FilterABC):
 
     def __init__(self, languages=None, id_method='langid', thresholds=None,
                  fasttext_model_path=None, langid_languages=None, cld2_options=None,
-                 lingua_mode=None, **kwargs):
+                 lingua_mode=None, heliport_options=None, **kwargs):
         super().__init__(**kwargs)
         if languages is None:
             raise ConfigurationError("A list of language codes needs to be defined")
@@ -326,6 +327,11 @@ class LanguageIDFilter(FilterABC):
         else:
             if lingua_mode:
                 raise ConfigurationError("lingua_mode is supported only by the method lingua")
+        if id_method == "heliport":
+            self.init_heliport(**{} if heliport_options is None else heliport_options)
+        else:
+            if heliport_options is not None:
+                raise ConfigurationError("heliport_options is supported only by the method heliport")
         # global options
         self.languages = languages
         self.id_method = id_method
@@ -367,6 +373,10 @@ class LanguageIDFilter(FilterABC):
             self.lingua_detector = from_languages.with_low_accuracy_mode().build()
         else:
             raise ConfigurationError(f"lingua mode '{lingua_mode}' is not supported.")
+
+    def init_heliport(self, ignore_confidence=False):
+        from heliport import Identifier
+        self.identifier = Identifier(ignore_confidence=ignore_confidence)
 
     def confidence(self, sentence: str, lan: str) -> float:
         """Return confidence of the identifier"""
@@ -415,6 +425,16 @@ class LanguageIDFilter(FilterABC):
             else:
                 liconf = confidence
             return liconf
+
+        if self.id_method == 'heliport':
+            iso_code_639_3 = self.identifier.identify(sentence)
+            if iso_code_639_3 == 'und':
+                # special label for too low confidence
+                return 0.5
+            lang = Lang(iso_code_639_3).pt1  # convert to ISO 639-1
+            if lang != lan:
+                return 0.0
+            return 1.0
 
         raise ValueError(f"Unknown language identification method '{self.id_method}'")