+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - support for selection of possible languages for `lingua` language detection
+- `letter_words_only` option for duplicate detection
 
 ### Fixed